Feat: display data user on home and update code display on profile, add useContext
