+++ src/context/user.jsx
@@ -1,0 +1,9 @@
+import { createContext, useContext } from "react";
+
+const UserContext = createContext();
+
+export function useUser() {
+  return useContext(UserContext);
+}
+
+export default UserContext;
